Fine, I'll do it myself.
So I've implemented a NO-BS togglable as-you-type autocompletion
that uses vim's internal autocompletion to do the job, and I've
even made two modes -- for normal completion and for code completion
from the tags file. Some changes are still yet to be made, as some
bindings are not ideal, but It's already quite good, so yeah.